refactor(modal dialog): Refactored modal dialog system with improved widget management and li...
# Project State

## Current Focus
Refactored modal dialog system with improved widget management and lifecycle tracking

## Context
This change improves the modal dialog system by:
1. Adding proper lifetime management to modal components
2. Fixing a potential buffer overflow in breadcrumb rendering
3. Aligning with recent widget lifecycle refactoring efforts

## Completed
- [x] Added lifetime parameters to `HelpOverlay` struct for proper modal management
- [x] Fixed breadcrumb rendering calculation to prevent potential buffer overflow
- [x] Updated widget imports to include `WidgetId` for better widget management

## In Progress
- [ ] No active work in progress for this commit

## Blockers
- None identified in this commit

## Next Steps
1. Verify modal dialog behavior in the demo examples
2. Ensure theme switching works correctly with the new modal implementation
---
category: refactor
scope: modal dialog
---

+++ tools/tui/dracon-terminal-engine/examples/theme_switcher.rs
@@ -628,7 +628,7 @@ impl Widget for WidgetDemoPanel {
         for (i, crumb) in crumbs.iter().enumerate() {
             let is_last = i == crumbs.len() - 1;
             if i > 0 {
-                let idx = (breadcrumb_row * area.width + x as u16) as usize;
+                let idx = breadcrumb_row as usize * area.width as usize + x as usize;
                 if idx < plane.cells.len() {
                     plane.cells[idx].char = '/';
                     plane.cells[idx].fg = theme.inactive_fg;
